api/requirements.txt
fastapi>=0.100.0
python-multipart>=0.0.6  # for file uploads
uvicorn[standard]>=0.23.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
pydantic>=2.0.0
